GoogleSearch + Additional Sites
Now with google custom search engine, please check if the search engine works on your system as it might not like the different ip address.

Added price scraping to amazon.ca and canadarobotix.com, adding additional sites is easy peasy but needs to also be added in the custom search engine which only I have access to.